Split core API source
